use root as default state
