+++ kahuna/package.json
@@ -35,8 +35,8 @@
       "uri-templates": "npm:uri-templates@0.1.5"
     },
     "devDependencies": {
-      "traceur": "github:jmcriffey/bower-traceur@0.0.91",
-      "traceur-runtime": "github:jmcriffey/bower-traceur-runtime@0.0.91"
+      "traceur": "github:jmcriffey/bower-traceur@0.0.92",
+      "traceur-runtime": "github:jmcriffey/bower-traceur-runtime@0.0.92"
     },
     "overrides": {
       "github:dbushell/Pikaday@1.3.2": {
