+++ kahuna/public/js/routes/index.js
@@ -1,11 +1,17 @@
 import angular from 'angular';
 
 import './search';
-import './results';
 import './image';
 
 export const routes = angular.module('gr.routes', [
     'gr.routes.search',
-    'gr.routes.results',
     'gr.routes.image'
 ]);
+
+routes.config(['$urlRouterProvider', '$locationProvider',
+               function($urlRouterProvider, $locationProvider) {
+
+    // Use real URLs (with History API) instead of hashbangs
+    $locationProvider.html5Mode({enabled: true, requireBase: false});
+    $urlRouterProvider.otherwise('/');
+}]);
