+++ kahuna/public/js/search/view.html
@@ -18,4 +18,5 @@
 
 <ui-view name="results"></ui-view>
 <ui-view name="panel"></ui-view>
+
 <dnd-uploader></dnd-uploader>
